requirements.txt
Flask~=3.1.0
peewee~=3.17.8
rich~=13.9.4
python-dotenv~=1.0.1
discord~=2.3.2
requests~=2.32.3
discord.py~=2.4.0
google~=3.0.0
protobuf~=5.29.3
Werkzeug~=3.1.3
brasilapy